add k8s and change order

+++ src/main/java/product_store/order/OrderModel.java
@@ -1,8 +1,10 @@
 package product_store.order;
 
+import java.math.BigDecimal;
+
 // Usado internamente para retornar a lista sem os itens
 public record OrderModel(
     String id,
     String date,
-    Float total
+    BigDecimal total
 ) {}
